Fix scaling issues with an ugly scale hammer

+++ src/ui/utils/screen_scale.rs
@@ -3,13 +3,13 @@ use serde::{Deserialize, Serialize};
 
 #[derive(Debug, Deserialize, Serialize)]
 pub struct ScreenScale {
-    pub scale: f64,
+    pub scale: f32,
 }
 
 impl ScreenScale {
     pub fn new(ctx: &mut Context) -> Self {
         ScreenScale {
-            scale: ctx.gfx.window().scale_factor(),
+            scale: ctx.gfx.window().scale_factor() as f32,
         }
     }
 }
